+++ src/pages/PaymentSuccess/index.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 import { useNavigate } from "react-router-dom";
-import Button from "../../components/Button";
+import Button from "../../components/common/Button";
 import SUCCESS from "/public/img/Success.png";
 
 interface CardProps {
